Enlarge your max input vars
Part of story #8332: ease mass update of mirroring

Starting from PHP 5.3.9, max_input_vars limit the number of variables
you can send to a php script. This affects mass update of repositories
when you have to update 1000 repositories / 10 mirrors.

Unfortunately, max_input_vars cannot be updated during execution time.
In order to limit the scope of this change, we only increase it at git
level.

If the limit (15000 ~= 1500 repositories / 10 mirrors) is reached, we
display a warning to the user.

Change-Id: Ic873dd00b7457565ecfe200bb3974faa4848a73e