Implements ldap_id reference
This task is part of story #8289 synchronize users in users groups

The route The route PUT /user_groups/:id/users is now able to
deal with references like {"ldap_id": 20001}

If the user is not found in LDAP, an error is thrown.

If the user is in LDAP but not yet in Tuleap database,
its account will be created automatically.

Change-Id: I6628e35aa2963e83ea1432f07b7501a8f324a87f