Introduce *insecure* option in email gateway
Part of story #6256: to be able to create an artifact by sending an email

What are the impacts for the site administrators?
-------------------------------------------------
The option sys_enable_reply_by_mail is moved from the local.inc to the
database. Starting from now, in order to configure email gateway,
administrators must go to: site admin » plugins » tracker.

A forgeupgrade will set the value in database accordingly to the value
of $sys_enable_reply_by_mail in the local.inc file currently used.
- if the variable is not in local.inc, then the feature was not used,
  therefore disable the feature in db.
- if the variable is set to 0, then the feature was not used, therefore
  disable the feature in db.
- if the variable is set to 1, then the feature is used with tokens,
  therefore activate the feature in 'token' mode in db.

If the variable is not cleared from the local.inc, then a warning is
displayed at the top of the config admin page.

What are the impacts for the end users?
---------------------------------------
- When the site admin disable the feature, no reply by mail is possible
  (tracker notifications to people are *not* sent with any special
  headers).
- When the site admin switches to token based email gateway, reply by
  mail is possible (tracker notifications to people are sent with
  special headers), like before.
- When the site admin switches to insecure email gateway, tracker admins
  can choose to activate for each tracker the feature (in tracker »
  administration » general settings). However please note that it is a
  work in progress, therefore artifact creation/reply by mail are not
  possible yet. Tracker notifications to people are *not* sent with any
  special headers.

  The tracker option is kept during export/import and is inherited from
  template.

Change-Id: I3bc5fce794c556dc8c2de81c127445e85e2a7496