story #6256: add the tracker email info in the report header
If the feature is activated for the tracker (and insecure emailgateway
is activated), then an envelope icon is displayed in the toolbar. When
the user clicks on it, a popover is displayed and show the tracker email
address used to create artifacts.

A button allow the user to copy the address directly to the clipboard. If
the clipboard is not accessible in the browser then no button, a link
is displayed instead, so that we gracefully support old browsers. As
IE Edge displays a ugly security box when we try to access the clipboard,
we don't offer the copy to clipboard possibility.

Change-Id: I626cadcb267ec99103dc8cd06ac0eee174cc9039