Fix request #8464: Check if values set for field list are possible
We need to clear stuff cached by Tracker_FromElement_Factory and
TrackerFactory in order to retreive the values for the list fields.

This issue appears while importing data with XML that leads to a complexe
usage of PHP objects. These factories shared data in memory that leads to this
misbehaviour.

So to fix this, we introduce a new method that clear all the caches for the instance
already in use in the code execution.

Change-Id: Id60ab8620861096fd1b19a55a7ed8987189bacaa