story #8289: Take into account the project_members group
We change the way to add and remove members. Instead of
deleting all the members and adding the members passed through
the references, we now delete only the members not passed through the
references and we add the users newly added through the references.

To be consistent with the Web UI :

* We can remove a project member that is suspended
* We cannot add a suspended user as a project member
* We cannot remove a user that is project admin

Change-Id: I32c4f67e1956d56ea402b298a0a346a5fd7ef026