request #8519: Add the possibility to build the packages and install a fresh instance
To start the build and installation just launch the new script:
$ tools/rpm/build_and_run_packages.sh --src-dir=/path/to/your/tuleap/sources

This is a first step which is far from perfect: at the end of the process,
you will get a shell on your fresh Tuleap install. You could access the instance
via the container IP address. When you quit the shell, the Tuleap instance and
build data are destroyed.

Also, we cannot fully clean the temporary Tuleap sources folder due to the way
the enalean/tuleap-buildsprms image do the build. Some changes must be done in this
image to not mess with files rights.

Change-Id: Iea8610f9e92133ccf665a0212b3e55c0c6809779