story #8431: see email headers for changesets created by email
When an artifact is created by email — this is possible only if you
1) activate insecure email gateway in site admin » plugins » tracker
2) activate the feature at the tracker level in tracker » admin »
settings — then the raw email + headers are stored in case of a future
audit.

This change introduces the ability to display those information. As this
is sensitive information, this is only available for site
administrators. The site admin just have to go to the artifact and click
on the envelope to see the raw mail in a new tab.

The information is displayed even if the email gateway is in Token mode
or disabled.

We convert also notification button behavior from PrototypeJS to jQuery.

Change-Id: Ib4d7b75c7f3da48f35fcba13e60df110c6a1dd18