Take into account news system references
This task is part of request #8443: System keywords don't generate the right cross reference when cross projects
The news system keyword is now available cross-project.
You can test it for instance with an SVN commit message.

Please note that while you can link to a forum with the news keyword, it
will not create a cross-reference. Only linking to actual news bytes
will create cross-references.

Change-Id: I9ab52d0885f6c5b38101d490ac85cd07463cd58a