request #8490 MediaWikiLanguageExtensionBundle not loaded
There is a bug in system_check event when MW is not used
and when a MW language is not set.

As the localsettings part that deals with it is not used in the MLEB activation context,
we add a check for this specific part to be skiped if necessary.

Change-Id: I56f4979e7313d7c0bd2d2c10d73be82b74ce132d