Generate user CSV file
This task is part of story #8528 create nonexistant user account at
import time

Introduce new script that generates a mapping file in order to know
which users in the users.xml file correspond to users in the import
server.

To generate the file, you have to do:
  src/utils/php-launcher.sh src/utils/generate_user_mapping_for_project_import.php \
    -u admin \
    -i archive.zip \
    -o users.csv

The generated file is like:

  name,action,comments
  jdoe,activate,"Comment"
  nterray,create,"Comment"
  yrossetto,map:,"Comment"
  ...

The comment gives to the user some information about actions he has to
do on users in order to be able to do the import.

For instance, he had to determinate which user map yrossetto:
yrossetto,map:yannis,"Comment"

The generated CSV file is not used yet in the import process.

For more information about the data generation process, please see the
flowchart[1].

Note:
[1] https://tuleap.net/plugins/tracker/?aid=8528&field=1246&func=show-attachment&attachment=921

Change-Id: I258546408ed3ed9e2946e4dcf7025dc1ac3abe1a