Check user CSV file
This task is part of story #8528 create nonexistant user account at import time

Once the user mapping file has been generated by
src/utils/generate_user_mapping_for_project_import.php, one should
update it to check and fix potential issues. For example:

  name,action,comments
  achille,activate,"Existing user achille must be activated (was [S])"

After investigations, it appears that achille user on the source
platform does not correspond to the suspended user achille, but
corresponds to jdoe user on the target platform. Therefore the mapping
should be updated:

  name,action,comments
  achille,map:jdoe,"achille is jdoe"

Note: The comments are here only for generation, they are ignored during
import time. They can be changed by hand to explain/remember choice
made.

They are some forbidden actions. For example, achille is a suspended user
in target platform, therefore one cannot choose to create it, the
following mapping file will be forbidden:

  name,action,comments
  achille,create,"…"

Here are the list of actions that are permitted based on the type of the
user:

  +----------------------+--------+----------+-----+
  | user types \ actions | create | activate | map |
  +----------------------+--------+----------+-----+
  | To Be Created        | ✓      | .        | ✓   |
  | To Be Activated      | .      | ✓        | ✓   |
  | To Be Mapped         | .      | .        | ✓   |
  | Already Existing     | .      | ✓        | ✓   |
  | Email Does Not Match | .      | .        | ✓   |
  +----------------------+--------+----------+-----+

As explained by the title of this commit message, no actions occur, only
a feedback is displayed on the console.

Change-Id: Iaadf046baf93c381150450a65e030ce3ad298fbe