request #8703 Kanban real time security to NodeJs server diffuse messages
Purposes:
- Reload kanban when NodeJS server detect that the JWT is expired
- Tuleap puts user rights on JWT
- Tuleap puts for an action (move or add card) who has the right to see
- For the moment we don't manage the rights where we need to know if the user is a submitter or the recipient
- NodeJS server checks the rights for client before sending a message

To test: Pull the git repository nodejs/tuleap-realtime on tuleap.net
Moreover to test the behavior when the JWT is expired:
You can load a client with a duration of the token at 30 minutes and an other client with a duration at 1 minute.
You can change it in JWTenerator.php file.

Change-Id: I598272b3b0da9065adc08be54dadeda218a913df