Create user when needed
This task is part of story #8528 create nonexistant user account at import time

If the mapping file contains users to be created, and the option
--force-create-all-users is given, then create those users. This
means that if those users are referenced in the project.xml with the
username then it will be used. However if it is referenced by its ldap
id then it will not be used yet.

The created user has a suspended status and a random password.

Change-Id: I7ce779cb0c250b5cbcfadf51a94ae30b5e7912a4