Use users.xml mapping while importing project
This task is part of story #8528 create nonexistant user account at import time

When you import a project archive, there are two files:
- one that contains data of the projects (user members, artifacts, …)
- one that contains all users that are part of those data

Until now, whenever you tried to import data with unknown users in your
platform, those users were converted to weird users (anonymous with
email = an id).

This is not anymore the case. Now users are correctly mapped, both in
project ugroups and trackers.

This development involved the modification of many parts that rely on
xml import:
- tracker creation from xml structure
- migration from tv3 to tv5
- copy of artifacts
- and of course import of project

*Note:* --force-create option is not anymore needed.

Change-Id: I197fa591236a8d310978e7b91067851d67772a22