request #8614: Project license removal
License is useless. We cannot see any relevant usage about this information. It
seems that trove cat is more suitable/flexible to handle the different licenses
of a company.

There is no objection from the community about this removal, see devel #33467.

Now license is not anymore part of the project registration, the site admin, nor
the project showdetails page.

Change-Id: Ia3885aa8447e0377084d9ce42ffb9a28a140a60f