story #8382: Planning V2 is now Planning
Planning V2 is out of the lab mode and become the default one.
Planning V1 is no longer accessible and redirect to Planning V2.

For now, Planning V1 is not fully removed from the codebase as we want to
able to produce a not too much painful patch to reactivate it in case of emergency.

Change-Id: I9f790638c0099f2dda21598d50adac1ef5abeef5