active action is now noop
This task is part of story #8528 create nonexistant user account at import time

Tests show that activate is not needed since data own by suspended users
can be imported and are still owned by same users.

Therefore we only indicate to the person in charge of the import that
the user is not alive.

+ add missing test that should have been submitted with
I197fa591236a8d310978e7b91067851d67772a22

Change-Id: I9ecc0c6e3d7deee3a98583f217390bfe08875cdc