request #8627: Site admin browsing of user list is broken when restricted to a project
Fix commit 872421b7b049ba5d57d97e83239f62e66eebef75

In the siteadmin userlist page, removed the column number of projects when
listing users related to a specific project (as opposed to listing all the
users of the platform).

This follows request #7715 which implementation broke the page with a user
visible error. The page tried to display data that wasn't provided by the
DAO.

Change-Id: I75bd68bb660cf4da8cc337d6dd1bb6bb9d5e3b60