Allow to create users in a given status
This task is part of story #8528: create nonexistant user account at import time

There are situations when users should not be created as suspended users
but as active or restricted ones.

Now the create action in mapping file accepts the status as argument.
List of possible statuses is [A, S, R]. Default status is S.

Example of mapping file:

  name,action,comments
  jawad,create:R,"…"

Change-Id: If728e22cd9ad4ac4e56c384c4bd00ee78fe603ea