Missing pagination in /milestones/:id/milestones
This is part of story #8380: have short response time when I load Planning view

Every routes that deal with collection were paginated in Tuleap except
this one.

This is not anymore the case and the route is now more than two times
faster (1.02s instead of 2.88s to load 10 paginated submilestones
instead of 50 not paginated).

Change-Id: I750c8fd4171e21c1532f64e1869a15fc3bdc08b9