request #8517 Stories created under Sprint Milestone on V2 Planning are not linked to their Release
Given the following AD plannings:

Release ------------ Sprints
|                       |
|                       |
`Epics  ----------------`Stories

If I create a User story in a Sprint with an Epic parent that is not
linked to any Release, then the story will be automatically linked
To the parent Release of the Sprint.

If the Epic is already linked to a Release or if the Story has no
parent, the actual behaviour does not change.

This link are also propagated to parent plannings.

Change-Id: I0b8fdfd1b41362999ae080ddf33db4a0931a8a35