Define SVN paths to trigger Jenkins job
This task is part of story #8671 trigger CI build on svn paths

First, we remove the star operator that can lead to some confusions.
If you want to trigger the job for all path, keep the textarea empty.

If you define one or more paths (one path per line in the textarea),
the job will be trigger only if the SVN commit impacts on one or more
paths defined by the user.

Change-Id: I144f37d84295236dc3032b1ae21732ce7d3506d4