request #8717: Improper user groups authorization verification for SVN ressources
We now check when a project is private if an user belongings to an usergroup
is also a project's member. If not, this user is not marked as a member of this
usergroup in the SVNAccessFile. This ensure that only project's member can access
the SVN repository when the project is private.

Change-Id: If3ff3c01f1ee356cf4dc65d713c9f3cee690ccc3