Define projects that can use SVN tokens
This task is part of story #8608: User can manage his own svn tokens

In the site admin section, if your svn_auth_type is not in modperl,
you will be able to define which project will be authorized to use
SVN tokens in addition of classic username + password.

A system_event will be created to rewrite the codendi_svnroot.conf file
according to these news changes.

For now, the revoke part is not yet implemented.

Change-Id: I2f74c0530718b4f81263831dc84bd86f0721652c