request #10948: Drop dependency to Zend Framework 1
Zend Framework 1 is not supported since September 2016, it is
time for Tuleap to abandon it.

This contribution removes the only remaining part of ZF1 in
Tuleap: Zend_Mail. Zend\Mail from the Zend Framework 2 is used
as a replacement since Tuleap and its plugins already have a
dependency on some parts of it.

Also, Codendi_Mail does not expose any objects from ZF anymore
to make future change of library easier and limit the risk of
breaking something in the process.

Change-Id: I232569112ed19ab5c1caa9b373ea78a8226393ab