request #8746 Project XML export fails if changeset value cannot be retrieved
I some case, the changeset value cannot be retrieved (no entry in DB due
to an error in the artifact life cycle).

In this case, the project export script is not able to deal with this
and this leads to a fatal error.

The solution here is to remove NULL values from the changeset_values
array.

Change-Id: I4fed473005ddd418c98d2ed665b3b40074f0cd80