request #8773: Allow plugins to display a link in the navbar
Completely refactor teh way navbar is built under FlamingParrot so that
a plugin can introduce itself.

Usage is:

Listen to the hook:

    $this->addHook(Event::NAVBAR_ITEMS);

Insert stuff in the collection. For example in the tracker plugin:

    public function navbar_items(array $params) {
        $params['items']->addItemAfterAnotherOne('project', new FlamingParrot_NavBarItemLinkPresenter(
            'trackers',
            strpos($params['request_uri'], '/plugins/tracker/?group_id=101') === 0,
            '/plugins/tracker/?group_id=101',
            'My Corporate Trackers'
        ));
    }

You can insert a link at the end of the collection (addItem) or after an existing item like in the example above.

Change-Id: I6c7aa67171d6f0929b07aa6912e0d632ef1f8a4a