request #8769: SVN paths in CI not displayed due to js error
jQuery('selector') always returns something. Therefore we cannot test if
the result is undefined to apply a behavior. We must test if the
returned collection is empty instead.

Furthermore select[name=form_parent] is too generic as a selector to
know if we are editing trove categories.

This commit brings back svn paths in continuous integration job
configuration and should not break trove cat edition.

Change-Id: Iece7c1f8b10233d019900b179446161e0092a52a