story #8637: have details only for one card
An icon is displayed on the card. When the users click on the icon, the
details of the card are displayed regardless of the current preferences.

The state for each card is not kept in user preferences.

If the user toggle compact/view mode for the kanban then all cards are
forced to be collapsed/expanded, but user can still expand them one
after the other.

When card is added in place, then its state respect user preferences.
The same for real-time: if nodejs server push a new card, it will be
expanded/collapsed accordingly to user prefs.

Change-Id: I08654f774a382c540305b71ef42264c289b9e080