Define default mirrors per project
This task is part of story #8734 have new git repositories automatically mirrored

A new tab is available in Git administration if at least one mirror
is defined. This tab is like the Git repository administration
view, but with only one tab "Mirroring".

Git administrator is able to select default mirrors for the new
repositories. But for now, the new repo are not yet
automatically mirrored on creation.

Change-Id: I3e7b029c234c8fe2af08be4bb7ab48a58d60d921