Purge old temporary attachments
This is part of story #8762: add attachment in modal

Now that the absolute number of temporary files are not anymore limited,
we must limit the time they live on the filesystem. Therefore, all
temporary files that are not updated during the last `$sys_file_deletion_delay`
days (defined in local.inc) are deleted during the system check event.

This requires to change the way TemporaryFileManager is built. We don't
rely anymore on the user in order to build the manager. This type of
object should only depend on infrastructure objects.

Change-Id: I5f33bbd02fe1f0e14130757f2b564866a337ee1b