request #8840 Remove usage of Bless
IE9 is no longer supported.

Only one message to warn users, we don't distinguish between
compatibility mode or deprecated browser.

Furthermore we plan to remind users about deprecation, therefore we
change the button and feedback message (even if the remind is not done
yet).

On the bonus side, we don't break anymore the whole layout for
MSIE != 9.

Change-Id: I8a3227f3b6d03e3d15544a49e12a4bebefcb1303