story #5894 update the artifact link data model
Accordingly to hierarchy of trackers, whenever we link two artifacts, we
now save the nature of the link. Only _is_child is saved for now.

For example if I create a task under a story, then in the
tracker_changeset_value_artifactlink, the nature column will say
'_is_child'. If there is no direct hierarchy (a task for an epic, or a
user story in a sprint) then the nature is NULL.

This commit takes into account the tracker hierachy updates (either
remove all
children or adding new one).

Have fun ;)

Change-Id: I78614705f0d6ae629153c4c8915e98dd0cc3697b