Edit .SVNAccessFile from web ui
Part of story #8813: Manage permissions

Backport from svn core the ability for project admin to define the
content of the .SVNAccessFile through the web interface.

.SVNAccessFile define in one repository does not impact other
repositories, including the repository of svn core.

Please note that the automatic modification of .SVNAccessFile
following the user groups creation/renaming will be done in a next
commit.

Change-Id: Idcc93b9b6f05ae32de5af3d13436893c8d5f8fd0