story #8631: Clean up the library loading and keep information about which provider is used
We keep this information in the state parameter. This variable is used to protect against
CSRF attacks and must not be guessable. To be able to pass the provider ID in the state
we sign it with a random key. When we got the state back from the OpenID Connect provider
we check its integrity with the random key we have stored in session.

Change-Id: I13e6be1f6a74e9a9ce57e998d94ffadde7315303