Introduce tabbed navigation in repository settings
Part of story #8813: Manage permissions

Like in git, repository settings is organized in sections. The existing
one is related to notifications. Prepare the groundwork to welcome a new
one in the upcoming changeset (access control).

It is now clearer that we are in a specific repository.

A link is missing to go back to the repository, however I suggest that
we add it when we will really work on the UX (see story #8817).

Change-Id: Ie404490dd86717cf4ae78621d37100249a1303e4