request #8921: Makefile should be self-documented
Add a new `help` target that displays the documentation of commonly used
targets of the Makefile.

Either run `make` or `make help`.

No need to document intermediate targets (like .env) or targets that are only
used in ci or in docker (like autoload). We prefer to document targets that are
used by developers on a day to day basis.

Change-Id: Id050e134dd86ffcf1305582a5084a42779594cf9