request #8932: Migration from tv3 to tv5 is not resilient
Good-enough mode is not good enough.

If tv3 data is not correctly formated, we can have a situation where it
is impossible to create the first changeset due to a lack of data. In my
case I have a comment that has been submitted before the initial
changeset (sic).

The current implementation throws an exception in that case, and catch it
inside the for loop that creates changesets. This is a problem since if
no initial changeset has been created, then subsequent changesets won't
be created (this currently leads to a fatal error).

My proposal is to create a fake changeset in that case.

Change-Id: If8a3b72d47d205a1849683280c76c0dc4e26514e