request #8947: Graphs on trackers are not resilient to data corruption
When the field used to build a pie chart or a bar chart is not a list
field, then we got a fatal error.

Worst, as the data is cached in session, updating manually the database
is not sufficient, we must force users to logout in order to clean the
session and retrieve the latest data.

The ideal would be to not have wrong data in the database, however until
we fix the root cause (due to tv3 - tv5 migration), we should not
display a blank page to the end user.

To reproduce the issue, update manually the settings of pie or bar
charts and give the id of a string field. (Do not forget that when you
add a chart, you must save the report before seeing stuff in database)
Clear the session (logout/login) and go back to the report: fatal error.

Apply the patch: no moar fatal error and a message is displayed to
indicates that the settings of the chast are wrong.

Change-Id: I9a3c51d48e255479054e6e8a88df98821e50f284