Create plugin reference
This task is part of story #8815 Manage cross references between svn
revisions & the world

This commit allows Tuleap to parse Subversion references
for the new Subversion plugin.
This kind of reference are like : svn #repo01/1

If you add such a reference in an Artifact comment for instance,
the reference will be clickable.

For now, it does not work cross-project.
Old Subersion reference must work.

If you use the Docker stuff for you development setup,
please use that trick to have working references:
https://gerrit.tuleap.net/#/c/5046/1/src/common/include/Codendi_HTMLPurifier.class.php

Change-Id: If8ecaa4051a94971cd8c7d92b7949b02550d9b74