Deal with crossproject
this is part of story #8815 Manage cross references between svn revisions & the world

To add a cross project reference,
add a ref in the svn commit message in a "project A" like art #123
The artifact #123 is in the "project B".

If you click on art #123 you should be redirect
into the tracker plugin of "project B".
In the artifact #123, you should have a link to svn commit of "project A"

Change-Id: I599afed91af8864fbbd67506a965ce3d27ebe39f