First step to restrict AL natures usage
Part of story #8833: show, import and manipulate links

ArtifactLink natures will provide some incompatibilities with tracker
hierarchy (and services that relies on this hierarchy, like agile
dashboard). In order to limit the scope of breaking changes, we must
activate this new feature on a per project basis.

To allow/disallow a project, site admin must go to admin » plugins »
tracker » artifact links nature.

As of today there aren't any impacts for the end user as this commit only
introduce the needed table for the administration. Teasing: in the
following changesets, you will have:

* projects that can use natures cannot use hierarchy and agile dashboard
* projects that use hierachy or agile dashboard cannot use nature
* projects that don't use natures don't have the new ArtifactLink view
  (group of tabs)

Change-Id: Id259ef69ae9ac8e8fb86a7dd7701659400edbf67