Create PATCH git/:id
This task is part of story #8887 migrate git to gerrit with REST route

The PATCH route deals with migrate_to_gerrit and disconnect_from_gerrit.

For the migration, if the request is OK, an event is launched to migrated the repository
(we kept the same behaviour as defined in web UI).

For the disconnect, we disconnect from gerrit the repository with
options:

* noop   => does nothing
* read   => update gerrit project into READ-ONLY mod
* delete => delete the gerrit project (throws an error if delete plugin
            not installed)

You can also disconnect a repository not migrated without error.

The lack of REST tests is due to system_event checking
(we don't know how to do this in REST test context)

Change-Id: I09c6ef75160e5691e474574a86991cf83510d464