Refactoring: move value saving in a dedicated object
Part of story #8833: show, import and manipulate links

This step is necessary to be able to save the nature of artifact links
in next changesets.

No changes for end user. Code is just moved around. This allows to
remove public visibility (used in tests) for some internal methods.

Things to test:
* Update classic artifact links (add, remove)
* Update artifact links in a tracker hierarchy context (in both
  direction: link a user story to its task or a task to its user story)
* AgileDashboard Scrum: add sub milestones, add children for backlog items, ...
* XML import
* REST POST/PUT artifacts/:id
* Copy artifact with children

Change-Id: I51c7d2c1e54d66debc5505b1360ea218ffb4fe39