request #8971: Avoid conflict with PHP 7 Error class
Fatal errors could be generated if a random number was generated
and after that a method of the Tuleap error class called.

This is due to the usage of the library php-paragonie-random-compat
introduced by commit 6697cfa343798c6ac75bddb9e800cd23b628b80e.
This library backports some parts of PHP 7 including the new class
Error.

Change-Id: I68f99039eb165e09134199242e16ce818fadb57f