Keep existing nature for artifact links
Part of story #8833: show, import and manipulate links

Given a Release tracker that is parent of a Sprint tracker,
And a release R1 has a sprint S1
When I edit only the title of the release,
Then the nature _is_child should be kept in db.

This bug is only visible in database. In order to test it:
0. Do not apply this patch
1. Link S1 and R1
2. Execute the sql statement:
   SELECT *
   FROM tracker_changeset_value_artifactlink
   ORDER BY changeset_value_id DESC LIMIT 10;
3. Observe that there is the nature _is_child at the top of the result
   for S1 linked artifact.
4. Update the R1 title (do not touch the ArtifactLink field)
5. Execute the sql statement: _is_child is lost.
6. Checkout/apply this patch.
7. Run forgeupgrade
8. Execute the sql statement: _is_child is here again.
9. Update the R1 title (do not touch the ArtifactLink field)
10. Execute the sql statement: _is_child is still here.

Change-Id: I570234a14b1cfb4a719749f89601c5d0f9e564d0