No burndown for projects with nature
Part of story #8833: show, import and manipulate links

If a project uses artifact links nature, then it cannot have a burndown
as the parent/child information is not complete yet. Instead a warning
is displayed both for the admin and the end user.

For projects that don't use nature, nothing changes.

We don't remove the possibility to create the field as it will imply
that we update accordingly the xml import, and we don't want to
introduce complexity that will be removed in a next release when the
parent/child of artifact will be complete.

Change-Id: Ie087be84f04e2b6c228dc86723281cdea27c2d3c