Fix error for live view of artifact links without nature
Part of story #8833: show, import and manipulate links

Enter an id of an artifact to link, do not select a nature, and click on
"Add artifacts": the artifact is displayed in the corresponding tracker
tab. There isn't any errors on the console.

Enter an id of an artifact to link, select a nature, and click on "Add
artifacts": the artifact is displayed in the corresponding tracker tab.
There isn't any errors in the console.

Change-Id: I099663b14b5317ed6583ec6295c0ad0af3ab56ac