Fix fatal error on /account/
Part of story #8633: customize openid providers

As administrator, define a provider github with the relevant icon.

As a logged in user, connect with github, and go to your preferences
page. You should not have any mustache error, and you should see the
icon next to the provider name in third party apps panel.

Change-Id: Ia429e4702c489232196caa4ac8fc02c3dfd3e9a1