Fix rendering issues of tabs
Part of story #8833: show, import and manipulate links

Browsers don't really like visibility:hidden with nav tabs: the latter
are not correcly rendered, they appear at the middle of the page and,
worst, may appear stacked on two lines (hello firefox).

The trick here is to use display:none only for ArtifactLink field.

"Only", because the layout relies on the visibility:hidden property to
display fields in columns. Using display:none in this situation would
make all colums on the same row to be resized when one field enters in
edition. Not sexy at all.

Maybe replacing a table-based layout by display:flex would solve
elegantly the problem however we don't have time to investigate on it
for this story.

Change-Id: I55e1910d59c2037ee818342c8e3056835388d23e