Allow one to edit nature of existing links
Part of story #8833: show, import and manipulate links

Edit the field ArtifactLink, go to a tracker tab (in a project that uses
nature), and change the nature of an existing link. Submit => the new
nature is saved.

You should test the A.D., trackers in project without natures, and every
modified field.

Change-Id: I481c1ecb89288b78530f9d48ce663cdda9a981c1